requirements.txt
django
channels
channels_redis
djangorestframework
daphne
django-cors-headers
dj-rest-auth[with_social]

